chore(refactor): move packageMetadataSchema to config/types (refs #47)
