+++ src/config/types.ts
@@ -2,6 +2,17 @@ import { z } from 'zod';
 
 /*
  * ============================================================================
+ * Package Metadata
+ * ============================================================================
+ */
+
+export const packageMetadataSchema = z.looseObject({
+  version: z.string().min(1),
+});
+
+
+/*
+ * ============================================================================
  * Namespace Filters
  * ============================================================================
  */
