+++ test/unit/types.test.ts
@@ -1,6 +1,6 @@
 import { describe, expect, it } from 'vitest';
 
-import { packageMetadataSchema } from '@/types';
+import { packageMetadataSchema } from '@/config/types';
 
 describe('packageMetadataSchema', () => {
   it('accepts a real package.json shape with additional keys', () => {
